rename collection.js to CollectionTab.tsx
