+++ src/window_main/CollectionTab.tsx
